requirements.txt
flask
yfinance
requests
beautifulsoup4
python-dotenv
lxml
pandas